refactor: moving RoutesContext from 'src\components\EditPage\temp_data\trip_reducer.jsx' to 'src\contexts\RoutesContext.jsx'
