+++ src/components/EditPage/Panel/PanelBody/PanelBody.jsx
@@ -6,17 +6,14 @@ import AutoScrollTabs from './AutoScrollTabs/AutoScrollTabs';
 import AddBtn from './AddBtn';
 import { postMaps, postDestinations } from '@/api/editPage.js';
 import { useAuth } from '@/contexts/AuthContext';
+import { routes_actions, useRoutesDispatch } from '@/contexts/RoutesContext';
 import {
   tripInfo_actions,
   useTripInfo,
   useTripInfoDispatch,
 } from '@/contexts/TripInfoContext';
 
-import {
-  routes_actions,
-  useRoutesDispatch,
-  usePlaceInfo,
-} from '../../temp_data/trip_reducer';
+import { usePlaceInfo } from '../../temp_data/trip_reducer';
 
 export default function PanelBody() {
   const [openForm, setOpenForm] = useState(false);
