+++ src/components/EditPage/temp_data/trip_reducer.jsx
@@ -3,40 +3,14 @@
 import moment from 'moment';
 import { useContext, createContext, useReducer } from 'react';
 
-const RoutesContext = createContext();
-const RoutesDispatchContext = createContext();
 const PlaceInfoContext = createContext();
 const PlaceInfoDispatchContext = createContext();
-
-export const routes_actions = {
-  SET_IS_Loaded: 'SET_IS_Loaded',
-  SET_ROUTES: 'SET_ROUTES', // 儲存交通路線資訊陣列
-};
 
 export const placeInfo_actions = {
   // GET_PLACE_INFO: 'GET_PLACE_INFO', // 取得搜尋景點的資訊
   SET_PLACE_INFO: 'SET_PLACE_INFO', // 暫存搜尋景點的資訊
   DELETE_PLACE_INFO: 'DELETE_PLACE_INFO', // 刪除搜尋景點的資訊
 };
-
-function routesReducer(routes, action) {
-  switch (action.type) {
-    case routes_actions.SET_IS_Loaded: {
-      const newRoutesState = JSON.parse(JSON.stringify(routes));
-      newRoutesState.isLoaded = action.payload;
-      return newRoutesState;
-    }
-    case routes_actions.SET_ROUTES: {
-      const newRoutesState = JSON.parse(JSON.stringify(routes));
-      const newRoutes = JSON.parse(JSON.stringify(action.payload));
-      newRoutesState.routes = newRoutes;
-      return newRoutesState;
-    }
-    default:
-      console.log('routes dispatch error');
-      break;
-  }
-}
 
 function placeInfoReducer(placeInfo, action) {
   switch (action.type) {
@@ -51,21 +25,6 @@ function placeInfoReducer(placeInfo, action) {
   }
 }
 
-export function RoutesProvider({ children }) {
-  const [routes, routesDispatch] = useReducer(routesReducer, {
-    isLoaded: false,
-    routes: [],
-  });
-
-  return (
-    <RoutesContext.Provider value={routes}>
-      <RoutesDispatchContext.Provider value={routesDispatch}>
-        {children}
-      </RoutesDispatchContext.Provider>
-    </RoutesContext.Provider>
-  );
-}
-
 export function PlaceInfoProvider({ children }) {
   const [placeInfo, placeInfoDispatch] = useReducer(placeInfoReducer, {});
 
@@ -76,14 +35,6 @@ export function PlaceInfoProvider({ children }) {
       </PlaceInfoDispatchContext.Provider>
     </PlaceInfoContext.Provider>
   );
-}
-
-export function useRoutes() {
-  return useContext(RoutesContext);
-}
-
-export function useRoutesDispatch() {
-  return useContext(RoutesDispatchContext);
 }
 
 export function usePlaceInfo() {
